add backofficeUser svc and delivery layer + modified userSvc + neccessary migrations

+++ service/userservice/profile.go
@@ -8,7 +8,7 @@ import (
 func (s Service) Profile(req param.ProfileRequest) (param.ProfileResponse, error) {
 	const op = "userservice.Profile"
 
-	user, err := s.repo.GetUserById(uint64(req.UserID))
+	user, err := s.repo.GetUserById(uint(req.UserID))
 	if err != nil {
 		//!  - use rich error to develope error handling from different layers
 		return param.ProfileResponse{}, richerror.New(op).WithErr(err).
